first stable versoin 1

+++ src/Helpers/FileMedia.php
@@ -5,51 +5,60 @@ namespace Waad\Media\Helpers;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
-abstract class FileMedia
+class FileMedia
 {
-    protected $model;
-    protected $file;
-    protected $index;
-    protected $destinationPath;
-    protected $result = [];
-    protected $disk;
-    protected $path;
+    public $model;
+    public $mediaClass;
+    public $file;
+    public $index;
+    public $result = [];
+    public $disk;
+    public $path;
+    public $baseNames = [];
 
-
-    public function __construct()
+    public function __construct($model)
     {
+        $this->model = $model;
+        $this->mediaClass = config('media.media_class');
         $this->disk = config('media.disk');
         $this->path = config('media.path');
     }
 
-    public function uploading()
+    // upload files and return information of files saved
+    public function uploading($files, $index = 1)
     {
-        if (! $this->file) {
+        $this->file = $files;
+        $this->index = $index;
+
+        if (is_null($this->file)) {
             return;
         }
         $is_list = $this->isList($this->file);
-        $is_list ? $this->uploadMany($this->file) : $this->uploadOne($this->file);
+        $is_list ? $this->uploadMany() : $this->uploadOne();
 
         return $this->result;
     }
 
-    protected function uploadOne($file)
+    // upload file
+    private function uploadOne()
     {
-        $this->store($file, $this->index);
+        return $this->storeFile($this->file, $this->index);
     }
 
-    protected function uploadMany($files)
+    // upload list files
+    private function uploadMany()
     {
         $i = $this->index;
-        foreach ($files as $file) {
-            $this->store($file, $i++);
+        foreach ($this->file as $file) {
+            $this->storeFile($file, $i++);
         }
     }
 
-    protected function store($file, $index)
+    // save file to storage
+    private function storeFile($file, $index)
     {
         $baseName = $this->randomNameFile($file);
-        Storage::disk($this->disk)->putFileAs($this->path,$file,$baseName);
+        Storage::disk($this->disk)->putFileAs($this->path, $file, $baseName);
         $fileName = $file->getClientOriginalName();
         $mime = $file->getClientMimeType();
         $file_size = filesize($file);
@@ -57,46 +66,81 @@ abstract class FileMedia
         $this->addResult($baseName, $fileName, $mime, $file_size, $index);
     }
 
-    public function delete()
+    // add to media DB after saved in storage
+    public function storeMedia()
     {
-        $is_list = $this->isList($this->file);
-        return $is_list ? $this->deleteMany($this->file) : $this->deleteOne($this->file);
+        foreach ($this->result as $information) {
+            $media = [
+                'base_name' => $information['base_name'],
+                'file_name' => $information['file_name'],
+                'index' => $information['index'],
+                'mime_type' => $information['mime_type'],
+                'file_size' => $information['file_size'],
+                'model_type' => get_class($this->model),
+                'model_id' => $this->model->getKey(),
+                'user_id' => $information['user_id'],
+            ];
+            $this->model->media()->create($media);
+            array_push($this->baseNames, $information['base_name']);
+        }
+
+        return $this->baseNames;
     }
 
-    protected function deleteOne($file)
+    // delete media and files
+    public function delete($files)
+    {
+        $medias = $files == null ? $this->model->media : $this->mediaClass->whereIn('id', $files)->get();
+        $this->file = $medias;
+        if (config('media.delete_file', false)) {
+            $is_list = $this->isList($this->file);
+            $is_list ? $this->deleteMany($this->file) : $this->deleteOne($this->file);
+        }
+        $this->file->delete();
+    }
+
+    // to method destory file
+    private function deleteOne($file)
     {
         return $this->destory($file);
     }
 
-    protected function deleteMany($files)
+    // to method destory list files
+    private function deleteMany($files)
     {
         foreach ($files as $file) {
             $this->destory($file);
         }
     }
 
-    protected function destory($file)
+    // delete file from storage if delete_file is true in config/media.php
+    private function destory($file)
     {
-        return Storage::disk($this->disk)->delete(sprintf("%s/%s",[$this->path, $file->base_name]));
+        return Storage::disk($this->disk)->delete(sprintf('%s/%s', $this->path, $file->base_name));
     }
 
-    public function addResult($baseName, $filename, $mime, $file_size, $index)
+    // add information file saved in result list
+    protected function addResult($baseName, $filename, $mime, $file_size, $index)
     {
+        $user = auth()->user();
         $this->result[] = [
-            'basename' => $baseName,
-            'filename' => $filename,
-            'mime' => $mime,
+            'base_name' => $baseName,
+            'file_name' => $filename,
+            'mime_type' => $mime,
             'file_size' => $file_size,
             'index' => $index,
+            'user_id' => is_null($user) ? null : $user->getKey(),
         ];
     }
 
-    protected function isList($file)
+    // check is list array or no
+    private function isList($file)
     {
         return is_array($file);
     }
 
-    protected function randomNameFile($file, $length = 16)
+    // create random string for name of file
+    private function randomNameFile($file, $length = 16)
     {
         return Str::random($length).time().'.'.$file->extension();
     }
